minor cleanup, move some shared label tree logic to its own file instead of schemaBasedCompletions
